% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.tipi.R
\name{tipi}
\alias{tipi}
\title{Calculate TIPI scores}
\usage{
tipi(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
tibble
}
\description{
Calculates TIPI scores.  Expects a data frame
formatted by \code{\link[expfactory:process_expfactory_survey]{expfactory::process_expfactory_survey}}, containing
responses from a single participant.
}
\keyword{TIPI}
